+++ packages/shared/src/index.ts
@@ -34,3 +34,4 @@ export * from './sect-war';
 export * from './sect-missions';
 export * from './sect-shop';
 export * from './sect-season';
+export * from './elemental';